+++ src/model/InfoBank.cpp
@@ -5,11 +5,11 @@ void InfoBank::add_subject(const std::string &str) {
     subjects.insert(st);
 }
 
-void InfoBank::add_group(const size_t group) {
+void InfoBank::add_group(const int group) {
     groups.insert(group);
 }
 
-void InfoBank::add_room(const size_t room) {
+void InfoBank::add_room(const int room) {
     rooms.insert(room);
 }
 
@@ -17,11 +17,11 @@ const std::set<std::string> &InfoBank::get_subjects() const {
     return subjects;
 }
 
-const std::set<size_t> &InfoBank::get_groups() const {
+const std::set<int> &InfoBank::get_groups() const {
     return groups;
 }
 
-const std::set<size_t> &InfoBank::get_rooms() const {
+const std::set<int> &InfoBank::get_rooms() const {
     return rooms;
 }
 
@@ -29,10 +29,10 @@ bool InfoBank::contains_subject(const std::string &str) const {
     return subjects.find(std::string(str)) != subjects.end();
 }
 
-bool InfoBank::contains_group(const size_t group) const {
+bool InfoBank::contains_group(const int group) const {
     return groups.find(group) != groups.end();
 }
 
-bool InfoBank::contains_room(const size_t room) const {
+bool InfoBank::contains_room(const int room) const {
     return rooms.find(room) != rooms.end();
 }
